Clarify materialized node storage model
### Motivation
- Make the distinction explicit between materialized node identity and cached node values so identity-level records cannot be confused with optional cached columns.
- Ensure correctness of merge, migration, invalidation, and pull semantics by making freshness/timestamps a total table over materialized identities and by introducing a clear `missing` freshness state.
- Prevent unsafe minting of `valid` proofs and remove any code/spec that treated `values.keys()` as the materialized-node registry.

### Description
- Spec change: rewrite the flag-based inverse-validity spec so `identifiers_keys_map` is the materialized-node registry, `values` is cached value storage, `freshness`/`timestamps` are total materialized-node tables, and `valid` is the inverse cached-value proof relation, and add `missing` to `Freshness`.
- Storage/API change: `listMaterializedNodes()` now enumerates identifiers from the identifier registry and a new `listCachedNodes()` enumerates `values.keys()`; helpers and comments were updated to use the vocabulary `materialized node` vs `cached node`.
- Validation/merge/migration change: `assertValidFinalMergeState`, migration logic, sync merge transfer/plan, and snapshot-migration scripts were updated to enforce the invariants (total freshness/timestamps over identifiers_keys_map, `values` ⊆ identities, `valid` only references cached nodes, clearing validity on cached-value deletion, and preserving createdAt on identity preservation).
- Behavioral change: invalidation updates `modifiedAt` and marks cached nodes `potentially-outdated` (missing nodes are left alone), relowering/merge may delete a cached value while preserving the materialized identifier (freshness=`missing`), and recomputation behavior updates freshness/timestamps/valid as specified by the model.

### Testing
- Ran static checks with `npm run static-analysis` and fixed lint/typing issues; static analysis succeeded.
- Exercised focused suites with `npx jest backend/tests/migrate_snapshot_to_flag_validity.test.js backend/tests/sync_merge.test.js --runInBand`; the updated suites pass after the changes.
- During iteration `npm test -- --runInBand` surfaced legacy timestamp expectations that were updated to match the new total-timestamps invariant; those expectations were adjusted and the focused regression suites now pass.

Identities live in the registry, values live in their cupboard — all checked and locked. @ottojung
